t2/producers/requirements.txt
peewee==3.17.0
psycopg2==2.9.5
psycopg2-binary==2.8.5
kafka-python
confluent-kafka
